requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-dotenv
pydantic
transformers
torch
accelerate
request